+++ packages/terra-icon/src/icon/IconMultipleResultsCorner.jsx
@@ -1,0 +1,19 @@
+/* eslint-disable */
+import React from 'react';
+import IconBase from '../IconBase';
+
+const SvgIcon = (customProps) => {
+  const attributes = Object.assign({}, customProps);
+
+  return (
+    <IconBase {...attributes}>
+      <path d="M0 0h48v48L22.29 22.29 0 0z" ></path>
+    </IconBase>
+  );
+};
+
+SvgIcon.displayName = "IconMultipleResultsCorner";
+SvgIcon.defaultProps = {"viewBox":"0 0 48 48","xmlns":"http://www.w3.org/2000/svg"};
+
+export default SvgIcon;
+/* eslint-enable */